requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
numpy>=1.18.5
scipy>=1.4.1
pandas>=1.4.2
netCDF4>=1.6.0
statistics>=1.0.3.5

# Plotting ------------------------------------
matplotlib>=3.2.2

# BD ------------------------------------------
sqlite3>=2.6.0

# Serialization -------------------------------
pyyaml>=6.0